+++ namukkun-project/src/Components/Home_Components/Home.js
@@ -1,18 +1,36 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
+import LoginModal from '../LoginModal_Components/LoginModal';
 
+//Home_Components의 컴포넌트들을 여기 불러오면 됨. 화면에 띄울 페이지
 function Home(){
+  const [showModal, setShowModal] = useState(false);
+
+  const toggleModal = () => {
+    setShowModal(!showModal);
+  };
 
   return(
     <div>
-      <ThemeDiv>처음입니다.</ThemeDiv>
+      <LoginButton onClick = {toggleModal}>
+        로그인
+      </LoginButton>
+      <LoginModal show = {showModal} onClose={toggleModal} />
     </div>
   );
 }
 
-const ThemeDiv = styled.div`
-  //theme 세팅 다 해놔서 props로 해당 형식으로 부르기만 하면 됨. style 폴더에 thmee.js 파일 참고
-  color: ${(props)=> props.theme.colors.secondary}; 
+const LoginButton = styled.button`
+  background: #007bff;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  cursor: pointer;
+  border-radius: 5px;
+
+  &:hover {
+    background: #0056b3;
+  }
 `;
 
 export default Home;
